Add completions
The zsh completion workings/syntax are largely cribbed and I don't
understand much about what it's doing. But it seems to work, so long as
I get zsh configured correctly (which was a bit of a pain).

The bash completions, on the other hand, were a complete pleasure to
write. Lots of utility functions from git.