integrate, finish, release: improve messages
Just learned about the `--quiet` flag to `git checkout`, which is nice
because I've been noticing how noisy it is to see messages like "Already
on 'develop'." or whatever.

Also improved the messaging around branch deletion on `release`, since
"has no remote for pushing" was not really clear at all.