Use git rev-parse consistently
In consulting gitcli(7), I realized that apparently the behavior of `git
rev-parse --pareopt` that I thought was pretty annoying (namely the
auto-parsing of the `-h` flag) is actually really consistent with the
rest of git. Live & learn.

So instead of hand-rolling my own usage messages & option parsing, may
as well bake in git-rev-parse style option parsing consistently. Most of
the commands are really boring on this front; about all git-rev-parse
buys us here is the ability to use a `--` to separate out args. Seems
reasonable enough, especially for scripting and whatnot. And at least
now I *believe* options are being parsed more consistently across the
board.