help: use `git help --man` instead of direct `man`
This takes advantage of git's man viewer configurations, making it more
consistent with the rest of the git suite. Sure are a lot of config
variables I've never touched, so it's hard to track down where all I'm
missing things...

The only thing I'm a little concerned about here is the algorithm for
converting `git help` input into a man page name, since it's not
actually documented in `git help help`, but the code[1] has been stable
for the past 9 years:

- if missing, equivalent to `git help git`
- if prefixed by "git", just use the input directly as the page name
- if it's a git command, prefix the page name by "git-"
- otherwise prefix the page name by "git" (no dash)

So I rely on the second line of logic, since my man pages have dashes in
their names. Could also pass it through a dashdash, technically: `git
help --man -- -topics`. But that seems uglier.

[1] https://github.com/git/git/blame/7668cbc60578f99a4c048f8f8f38787930b8147b/builtin/help.c#L294-L304