reintegrate: fix links
Seems that since I last jotted down the URL to the git-rebase.sh &
git-rebase--interactive.sh scripts, the folks over on Git have been
working to reimplement git-rebase in C. Kudos to them, but I'm not yet
basing my implementation on their C code, so I fixed the links in my
comments. Credit where it's due and all that.